Move the postproc2 image and subject CLIs to the CLI module. Also move click command constants from config to the click module.
